requirements.txt
Pillow==8.4.0
pystyle==0.6
